online_shop/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.5
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-oauth-toolkit==2.2.0
django-payments==1.0.0
django-phonenumber-field==7.0.2
django-rest-authtoken==2.1.4
django-timezone-field==5.0
djangorestframework==3.14.0
drf-social-oauth2==1.2.1
drf-writable-nested==0.7.0
ecdsa==0.18.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
isort==5.12.0
jwcrypto==1.4.2
kombu==5.2.4
mccabe==0.7.0
mercadopago==2.2.0
oauthlib==3.2.2
phonenumbers==8.13.5
phonenumberslite==8.13.6
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stripe==5.1.1
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
wrapt==1.14.1
